# packages/shared/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "CharacterAppearance": {
        "additionalProperties": false,
        "properties": {
          "age_visual": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Age Visual",
            "type": "string"
          },
          "body": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "face": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Face",
            "type": "string"
          },
          "hair": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Hair",
            "type": "string"
          },
          "skin": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Skin",
            "type": "string"
          }
        },
        "required": [
          "age_visual",
          "face",
          "hair",
          "skin",
          "body"
        ],
        "title": "CharacterAppearance",
        "type": "object"
      },
      "CharacterClothing": {
        "additionalProperties": false,
        "properties": {
          "accessories": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Accessories",
            "type": "string"
          },
          "bottom": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Bottom",
            "type": "string"
          },
          "shoes": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Shoes",
            "type": "string"
          },
          "top": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Top",
            "type": "string"
          }
        },
        "required": [
          "top",
          "bottom",
          "shoes",
          "accessories"
        ],
        "title": "CharacterClothing",
        "type": "object"
      },
...
      "CharacterResponse": {
        "additionalProperties": false,
        "properties": {
          "appearance": {
            "$ref": "#/components/schemas/CharacterAppearance"
          },
          "character_id": {
            "title": "Character Id",
            "type": "string"
          },
          "clothing": {
            "$ref": "#/components/schemas/CharacterClothing"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "distinctive_features": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distinctive Features"
          },
          "master_description": {
            "title": "Master Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "personality_traits": {
            "items": {
              "type": "string"
            },
            "title": "Personality Traits",
            "type": "array"
          },
          "visual_style_notes": {
            "title": "Visual Style Notes",
            "type": "string"
          }
        },
        "required": [
          "character_id",
          "name",
          "master_description",
          "appearance",
          "clothing",
          "personality_traits",
          "visual_style_notes",
          "created_at"
        ],
        "title": "CharacterResponse",
        "type": "object"
      },
...
      "ConsentGrantRequest": {
        "properties": {
          "consent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Version"
          },
          "data_processing": {
            "default": false,
            "title": "Data Processing",
            "type": "boolean"
          },
          "photos": {
            "default": false,
            "title": "Photos",
            "type": "boolean"
          },
          "privacy": {
            "default": false,
            "title": "Privacy",
            "type": "boolean"
          }
        },
        "title": "ConsentGrantRequest",
        "type": "object"
      },
      "ConsentResponse": {
        "properties": {
          "consent_version": {
            "title": "Consent Version",
            "type": "string"
          },
          "data_processing": {
            "title": "Data Processing",
            "type": "boolean"
          },
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "photos": {
            "title": "Photos",
            "type": "boolean"
          },
          "privacy": {
            "title": "Privacy",
            "type": "boolean"
          },
          "revoked": {
            "title": "Revoked",
            "type": "boolean"
          }
        },
        "required": [
          "granted",
          "privacy",
          "photos",
          "data_processing",
          "consent_version",
          "revoked"
        ],
        "title": "ConsentResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "email": "support@aistorybook.com",
      "name": "AI Story Book Team"
    },
    "description": "\n# AI Story Book API\n\nAI 기반 맞춤형 동화책 생성 API입니다.\n\n## 주요 기능\n\n* **책 생성**: 주제, 연령대, 스타일을 입력하면 AI가 동화책을 생성합니다\n* **캐릭터 관리**: 사진에서 캐릭터를 추출하거나 직접 생성할 수 있습니다\n* **시리즈 생성**: 같은 캐릭터로 연속된 이야기를 만들 수 있습니다\n* **PDF/오디오 내보내기**: 완성된 책을 PDF나 오디오로 내보낼 수 있습니다\n\n## 인증\n\n모든 API는 `X-User-Key` 헤더가 필요합니다.\n중복 요청 방지를 위해 `X-Idempotency-Key` 헤더 사용을 권장합니다.\n\n## Rate Limiting\n\n- 기본: 분당 10회 요청 제한\n- 초과 시 429 Too Many Requests 응답\n\n## 크레딧 시스템\n\n책 1권 생성에 크레딧 1개가 소모됩니다.\n    ",
    "license": {
      "name": "MIT"
    },
    "title": "AI Story Book API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/characters/{character_id}": {
      "delete": {
        "description": "캐릭터 삭제",
        "operationId": "delete_character_v1_characters__character_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          },
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Character",
        "tags": [
          "Characters"
        ]
      },
      "get": {
        "description": "캐릭터 상세 조회",
        "operationId": "get_character_v1_characters__character_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          },
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Character",
        "tags": [
          "Characters"
        ]
      }
    },
...
    "/v1/consent": {
      "get": {
        "description": "현재 유효한 보호자 동의 상태 조회.",
        "operationId": "get_consent_v1_consent_get",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Consent",
        "tags": [
          "Consent"
        ]
      },
      "post": {
        "description": "보호자 동의 기록. granted 는 필수 동의(개인정보+데이터 처리) 충족 시 True.\n\n유저당 비철회 동의 행은 최대 1개가 되도록, 새 동의 시 기존 비철회 행을 폐기(supersede)한다.",
        "operationId": "grant_consent_v1_consent_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grant Consent",
        "tags": [
          "Consent"
        ]
      }
    },
    "/v1/consent/revoke": {
      "post": {
        "description": "보호자 동의 철회 + 수집된 아동 사진/그림 파생 캐릭터 파기(PIPA/COPPA 철회 의무).\n\n이후 사진 기반 기능은 다시 차단되고, from_photo 캐릭터 레코드·원본 사진(스토리지)은 삭제된다.",
        "operationId": "revoke_consent_v1_consent_revoke_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Consent",
        "tags": [
          "Consent"
        ]
      }
    },
...
  },
...
}